Move the interfaces definition of the table inside the table component
- Create the interfaces definition inside the table component
- Update all the references
