+++ src/components/table/data.entity.ts
