+++ src/components/table/index.ts
@@ -1,1 +1,3 @@
 export * from './table.component';
+export * from './data.entity';
+export * from './config.entity';
