+++ src/list-people/entities/index.ts
@@ -1,3 +1,1 @@
-export * from './data-table.entity';
-export * from './config-table.entity';
 export * from './page-limit.entity';
